Add persistence functionality.

+++ src/main/java/net/ripe/rpki/ta/persistence/TAPersistence.java
@@ -1,47 +1,72 @@
 package net.ripe.rpki.ta.persistence;
 
+/*-
+ * ========================LICENSE_START=================================
+ * RIPE NCC Trust Anchor
+ * -
+ * Copyright (C) 2017 RIPE NCC
+ * -
+ * Redistribution and use in source and binary forms, with or without modification,
+ * are permitted provided that the following conditions are met:
+ *
+ * 1. Redistributions of source code must retain the above copyright notice, this
+ *    list of conditions and the following disclaimer.
+ *
+ * 2. Redistributions in binary form must reproduce the above copyright notice,
+ *    this list of conditions and the following disclaimer in the documentation
+ *    and/or other materials provided with the distribution.
+ *
+ * 3. Neither the name of the RIPE NCC nor the names of its contributors
+ *    may be used to endorse or promote products derived from this software without
+ *    specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+ * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED.
+ * IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
+ * INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
+ * BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
+ * LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE
+ * OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED
+ * OF THE POSSIBILITY OF SUCH DAMAGE.
+ * =========================LICENSE_END==================================
+ */
+
 import com.google.common.base.Charsets;
 import com.google.common.base.Preconditions;
 import com.google.common.base.Strings;
 import com.google.common.io.Files;
-import net.ripe.rpki.commons.crypto.util.KeyPairFactory;
-import net.ripe.rpki.ta.cas.TrustAnchor;
+import net.ripe.rpki.ta.config.Config;
 import org.apache.log4j.Logger;
 
 import java.io.File;
 import java.io.IOException;
-import java.nio.charset.Charset;
 
-public class TrustAnchorPersistenceHandler implements PersistenceHandler<TrustAnchor> {
+public class TAPersistence {
 
-    private static final Logger LOG = Logger.getLogger(TrustAnchorPersistenceHandler.class);
+    private static final Logger LOG = Logger.getLogger(TAPersistence.class);
 
-    static final String TRUST_ANCHOR_FILENAME = "ta";
-    static final String TRUST_ANCHOR_FILE_EXT = "xml";
-    static final Charset IO_CHARSET = Charsets.UTF_8;
-
-    private TrustAnchorSerializer serializer = new TrustAnchorSerializer();
+    private static final String TRUST_ANCHOR_FILENAME = "ta";
+    private static final String TRUST_ANCHOR_FILE_EXT = "xml";
 
     private final File persistenceDirectory;
     private final File trustAnchorFile;
 
-    private final KeyPairFactory keyPairFactory;
-
-    public TrustAnchorPersistenceHandler(String persistenceDirectory, KeyPairFactory keyPairFactory) {
-        this.persistenceDirectory = new File(persistenceDirectory);
-        if (! this.persistenceDirectory.exists()) this.persistenceDirectory.mkdirs();
+    public TAPersistence(final Config config) {
+        this.persistenceDirectory = new File(config.getPersistentStorageDir());
+        if (!this.persistenceDirectory.exists()) this.persistenceDirectory.mkdirs();
         Preconditions.checkArgument(this.persistenceDirectory.isDirectory(), "Can't create directory: " + persistenceDirectory);
         this.trustAnchorFile = new File(this.persistenceDirectory, TRUST_ANCHOR_FILENAME + "." + TRUST_ANCHOR_FILE_EXT);
-        this.keyPairFactory = keyPairFactory;
     }
 
-    @Override
-    public void save(TrustAnchor trustAnchor) throws IOException {
+    public void save(String xml) throws Exception {
+        if (trustAnchorFile.exists()) {
+            throw new Exception("File " + trustAnchorFile.getAbsolutePath() + " already exists");
+        }
         final File tempFile = File.createTempFile(Strings.padStart(TRUST_ANCHOR_FILENAME, 3, '_'), TRUST_ANCHOR_FILE_EXT, persistenceDirectory);
         try {
-            String serializedXml = serializer.serialize(trustAnchor);
-            Files.write(serializedXml, tempFile, IO_CHARSET);
-            backupOldFileIfItExists();
+            Files.write(xml, tempFile, Charsets.UTF_8);
             Files.move(tempFile, trustAnchorFile);
             LOG.info("Trust Anchor written to: " + trustAnchorFile);
         } finally {
@@ -49,18 +74,7 @@ public class TrustAnchorPersistenceHandler implements PersistenceHandler<TrustAn
         }
     }
 
-    private void backupOldFileIfItExists() throws IOException {
-        if (trustAnchorFile.exists()) {
-            File destTaFile = new File(persistenceDirectory, TRUST_ANCHOR_FILENAME + "-" + System.currentTimeMillis() + "." + TRUST_ANCHOR_FILE_EXT);
-            Files.move(trustAnchorFile, destTaFile);
-        }
-    }
-
-    @Override
-    public TrustAnchor load() throws IOException {
-        TrustAnchor trustAnchor = serializer.deserialize(Files.toString(trustAnchorFile, IO_CHARSET));
-        trustAnchor.open(keyPairFactory);
-        LOG.info("Trust Anchor read from: " + trustAnchorFile);
-        return trustAnchor;
+    public String load() throws IOException {
+        return Files.toString(trustAnchorFile, Charsets.UTF_8);
     }
 }
